requirements.txt
fastapi
keras
numpy
opencv-python
python-multipart
Pillow